Check that a new func schema has outputs != nil
A prior change (#20) added a check that the field `.Outputs` on the old
schema is not nil, when comparing with a new schema. This suggests it's
possible for the outputs on the _new_ func schema to be nil, so check
for that too. This avoids a possible nil pointer access
(`newFunc.Outputs.Properties`), and gives an accurate violation message.