Change the output format to snake_case
This PR changes the output format of `schema-tools stats` to snake_case. The change was
recommended by @sicarul as snake_case better preserves capitalization in our analytics
pipeline.

I have been unable to find any programmatic dependencies on the output of this function in
Pulumi, so this change should not break downstream users.